+++ packages/media/media-store/models/auth-query-parameters.d.ts
@@ -1,0 +1,11 @@
+import { Auth } from './auth';
+export interface ClientBasedQueryParameters {
+    readonly client: string;
+    readonly token: string;
+}
+export interface AsapBasedQueryParameters {
+    readonly issuer: string;
+    readonly token: string;
+}
+export declare type AuthQueryParameters = ClientBasedQueryParameters | AsapBasedQueryParameters;
+export declare function mapAuthToQueryParameters(auth: Auth): AuthQueryParameters;