+++ packages/media/media-store/models/auth.js
@@ -1,0 +1,7 @@
+export function isClientBasedAuth(auth) {
+    return !!auth.clientId;
+}
+export function isAsapBasedAuth(auth) {
+    return !!auth.asapIssuer;
+}
+//# sourceMappingURL=auth.js.map