+++ packages/media/media-store/models/media.js
@@ -1,0 +1,4 @@
+export var isPreviewableType = function (type) {
+    return ['audio', 'video', 'image'].indexOf(type) > -1;
+};
+//# sourceMappingURL=media.js.map