+++ packages/media/media-store/upload-controller.d.ts
@@ -1,0 +1,7 @@
+export declare type AbortFunction = () => void;
+export declare class UploadController {
+    abortFunction?: AbortFunction;
+    constructor();
+    setAbort(abortFunction: AbortFunction): void;
+    abort(): void;
+}