+++ packages/media/media-store/utils/hashing/hasherCreator.d.ts
@@ -1,0 +1,3 @@
+import { Hasher } from './hasher';
+export declare const destroyHasher: () => any;
+export declare const createHasher: () => Promise<Hasher>;