+++ packages/media/media-store/utils/hashing/simpleHasher.d.ts
@@ -1,0 +1,4 @@
+import { Hasher } from './hasher';
+export declare class SimpleHasher implements Hasher {
+    hash(blob: Blob): Promise<string>;
+}