+++ packages/media/media-store/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-store
 
+## 0.1.4
+- [patch] :
+
+  - Built with ts, changed package with version
+
 ## 0.1.3
 - [patch] :